% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhReg.R
\name{fourier_K}
\alias{fourier_K}
\title{Fourier K}
\usage{
fourier_K(Fit)
}
\arguments{
\item{Fit}{Model built using dhr function}
}
\value{
optimal value of K used in dhr function
}
\description{
function to get best value of K used in dhr function
}
\examples{
\donttest{
Data1 <- runif(runif(200,100,1000))#To generate random number for example
Data_ts <- ts(Data1)
M <- dhr(Data=Data_ts,XREG=NULL,Range=list(1:2,1),Frequency=c(24,168),Criteria="aicc")
fourier_K(M)
}
}
